backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.7.1
Flask-Cors==6.0.1
Flask-Limiter==3.5.0
marshmallow==3.20.1
python-dotenv==1.1.1
PyMySQL==1.1.2
cryptography==46.0.3
bcrypt==5.0.0

# Dependencias para chatbot y parseo de archivos
google-generativeai==0.8.3
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5